# feat: reduce map access