I reduced map access to improve newGraph's performance.

```
name        old time/op    new time/op    delta
NewGraph-8     1.37s ± 4%     1.17s ± 5%  -14.89%  (p=0.000 n=10+10)

name        old alloc/op   new alloc/op   delta
NewGraph-8    47.8MB ± 3%    46.9MB ± 2%     ~     (p=0.075 n=10+10)

name        old allocs/op  new allocs/op  delta
NewGraph-8      555k ± 3%      548k ± 3%     ~     (p=0.190 n=10+10)
```